+++ slider/src/brain/commands.cpp
@@ -1,5 +1,5 @@
 #include "commands.h"
-#include "settings.h"
+#include "src/settings/settings.h"
 #include "src/core/utils/mathUtils.h"
 
 using namespace Core;
