+++ slider/src/settings/settings.cpp
