+++ slider/src/settings/settings.h
